refactor: read  page for omegaquotes
