+++ src/actions/omegaquotes/read.ts
@@ -1,15 +1,14 @@
 'use server'
-
-import prisma from '@/prisma'
-import { createActionError, createPrismaActionError } from '@/actions/error'
+import { createActionError } from '@/actions/error'
 import { getUser } from '@/auth/user'
 import type { ActionReturn, ReadPageInput } from '@/actions/Types'
-import type { OmegaquoteFiltered } from './Types'
+import type { OmegaquoteFiltered } from '@/server/omegaquotes/Types'
+import { readQuotesPage } from '@/server/omegaquotes/read'
 
-export async function readQuotesPage<const PageSize extends number>(
-    { page }: ReadPageInput<PageSize>
+export async function readQuotesPageAction<const PageSize extends number>(
+    ReadPageInput: ReadPageInput<PageSize>
 ): Promise<ActionReturn<OmegaquoteFiltered[]>> {
-    // REFACTOR when new permission system is working
+    //TODO:  REFACTOR when new permission system is working
     const { status } = await getUser({
         requiredPermissions: ['OMEGAQUOTES_READ']
     })
@@ -18,24 +17,5 @@ export async function readQuotesPage<const PageSize extends number>(
         return createActionError(status)
     }
 
-    try {
-        const { page: pageNumber, pageSize } = page
-        const results = await prisma.omegaQuote.findMany({
-            orderBy: {
-                timestamp: 'desc',
-            },
-            skip: pageNumber * pageSize,
-            take: pageSize,
-            select: {
-                id: true,
-                author: true,
-                quote: true,
-                timestamp: true,
-            },
-        })
-
-        return { success: true, data: results }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    return await readQuotesPage(ReadPageInput)
 }
