+++ src/server/omegaquotes/Types.ts
@@ -1,3 +1,4 @@
 import type { OmegaQuote } from '@prisma/client'
+import { omegaQuoteFieldsToExpose } from './CofigVars'
 
-export type OmegaquoteFiltered = Pick<OmegaQuote, 'id' | 'author' | 'quote' | 'timestamp'>
+export type OmegaquoteFiltered = Pick<OmegaQuote, typeof omegaQuoteFieldsToExpose[number]>
